Move RunRC, DeleteRC, and ListPods into the utility module and refactor to be support functions #7234
